backend/auth-service/requirements.txt
fastapi==0.110.0
pydantic==2.6.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
uvicorn[standard]==0.27.1
httpx==0.27.0